Modufy fetchLambdaInfo

+++ pkg/madmin/info-commands.go
@@ -408,12 +408,12 @@ type Services struct {
 
 // Buckets contains the number of buckets
 type Buckets struct {
-	Count int `json:"count,omitempty"`
+	Count uint64 `json:"count,omitempty"`
 }
 
 // Objects contains the number of objects
 type Objects struct {
-	Count int `json:"count,omitempty"`
+	Count uint64 `json:"count,omitempty"`
 }
 
 // Usage contains the tottal size used
